remove deprecated code paths
