--- src/http/_session-read.js
@@ -1,10 +1,0 @@
-let jwe = require('./providers/jwe')
-let ddb = require('./providers/ddb')
-
-module.exports = function read(request, callback) {
-
-  if (process.env.SESSION_TABLE_NAME === 'jwe')
-    return jwe.read(request)
-
-  return ddb.read(request, callback)
-}
