--- src/http/_session-write.js
@@ -1,10 +1,0 @@
-let jwe = require('./providers/jwe')
-let ddb = require('./providers/ddb')
-
-module.exports = function write(params, callback) {
-
-  if (process.env.SESSION_TABLE_NAME === 'jwe')
-    return jwe.write(params)
-
-  return ddb.write(params, callback)
-}
