+++ src/http/helpers/static.js
