--- src/http/session/_week-from-now.js
@@ -1,4 +1,0 @@
-// epoch + one week in seconds
-module.exports = function _weekFromNow() {
-  return Date.now() + 604800
-}
