--- src/http/session/index.js
@@ -1,7 +1,0 @@
-var client = require('./client')
-var init = require('./_init')
-
-module.exports = {
-  init,
-  client,
-}
