--- src/http/session/init.js
@@ -1,19 +1,0 @@
-var db = require('./_get-dynamo-instance')
-
-module.exports = function init(name, callback) {
-  db.createTable({
-    TableName: name,
-    KeySchema: [{
-      AttributeName: '_idx',
-      KeyType: 'HASH'
-    }],
-    AttributeDefinitions: [{
-      AttributeName: '_idx',
-      AttributeType: 'S'
-    }],
-    ProvisionedThroughput: {
-      ReadCapacityUnits: 5,
-      WriteCapacityUnits: 5
-    }
-  }, callback)
-}
