+++ src/http/session/providers/ddb/_week-from-now.js
@@ -1,0 +1,6 @@
+// epoch + one week IN SECONDS
+// - Date.now() returns ms
+// - DynamoDB TTL is in seconds
+module.exports = function _weekFromNow() {
+  return (Date.now() / 1000) + 604800
+}
