--- src/http/validate/_status-codes.js
@@ -1,1 +1,0 @@
-module.exports = [400, 403, 404, 406, 409, 415, 500]
