+++ test/02-session-test.js
@@ -1,100 +1,9 @@
-var session = require('../src/http/session')
-var test = require('tape')
-var dynalite = require('dynalite')
-var server, sesh
-var name = 'test-sessions'
+var read = require('../src/http/session/read')
+var write = require('../src/http/session/read')
+let test= require('tape')
 
-test('can startup dynalite server', t=> {
+test('http.session', t=> {
   t.plan(2)
-  server = dynalite({createTableMs:0}).listen(5000, err=> {
-    if (err) {
-      t.fail(err)
-    }
-    else {
-      t.ok(true, 'Dynalite started on port 5000')
-      t.ok(session, 'exists in scope')
-    }
-  })
-})
-
-test('can init', t=> {
-  t.plan(1)
-  session.init('test-sessions', function _init(err) {
-    if (err) {
-      t.fail(err)
-    }
-    else {
-      t.ok(true, 'initialized')
-    }
-  })
-})
-
-test('can get a session client', t=> {
-  t.plan(1)
-  session = session.client(name)
-  t.ok(session, 'got a session client')
-})
-
-test('can create', t=> {
-  t.plan(1)
-  session._create({foo:1}, function _create(err, data) {
-    if (err) {
-      t.fail(err)
-    }
-    else {
-      t.ok(data, 'created')
-      sesh = data
-      console.log(data)
-    }
-  })
-})
-
-test('can find', t=> {
-  t.plan(1)
-  session._find(sesh._idx, function _find(err, data) {
-    if (err) {
-      t.fail(err)
-    }
-    else {
-      t.ok(data, 'found')
-      console.log(data)
-    }
-  })
-})
-
-test('can update', t=> {
-  t.plan(1)
-  sesh.b = 2
-  session._update(sesh, function _update(err, data) {
-    if (err) {
-      t.fail(err)
-    }
-    else {
-      t.ok(data, 'updated')
-      sesh = data
-      console.log(data)
-    }
-  })
-})
-
-test('find on a dead key returns fresh key', t=> {
-  t.plan(2)
-  var deadkey = 'A8RSJpccsfGzR8QtH43kr4RV'
-  session._find(deadkey, function _find(err, data) {
-    if (err) {
-      t.fail(err)
-    }
-    else {
-      t.ok(data, 'found')
-      t.ok(deadkey != data._idx, 'gotta fresh _idx')
-      console.log({deadkey})
-      console.log(data)
-    }
-  })
-})
-
-test('can shutdown', t=> {
-  t.plan(1)
-  server.close()
-  t.ok(true, 'server closed')
+  t.ok(read, 'http.session.read')
+  t.ok(write, 'http.session.write')
 })
